+++ src/main/java/com/mingles/metamingle/interactivemovie/command/infrastructure/service/ApiShortFormService.java
@@ -1,15 +1,10 @@
 package com.mingles.metamingle.interactivemovie.command.infrastructure.service;
 
-import com.mingles.metamingle.auth.JwtTokenProvider;
 import com.mingles.metamingle.shortform.command.application.dto.response.CreateShortFormResponse;
-import com.mingles.metamingle.shortform.command.application.dto.response.UploadVideo;
 import com.mingles.metamingle.shortform.command.application.service.ShortFormFirebaseService;
-import com.mingles.metamingle.shortform.command.domain.aggregate.entity.ShortForm;
 import lombok.RequiredArgsConstructor;
 import org.jcodec.api.JCodecException;
 import org.springframework.stereotype.Service;
-import org.springframework.web.bind.annotation.RequestHeader;
-import org.springframework.web.multipart.MultipartFile;
 
 import java.io.IOException;
