+++ src/main/java/com/mingles/metamingle/movielike/command/application/controller/InteractiveMovieLIkeCommandController.java
@@ -1,7 +1,7 @@
 package com.mingles.metamingle.movielike.command.application.controller;
 
-import com.mingles.metamingle.auth.JwtTokenProvider;
-import com.mingles.metamingle.common.ApiResponse;
+import com.mingles.metamingle.global.auth.JwtTokenProvider;
+import com.mingles.metamingle.global.common.ApiResponse;
 import com.mingles.metamingle.movielike.command.application.dto.response.CreateInteractiveMovieLikeResponse;
 import com.mingles.metamingle.movielike.command.application.service.InteractiveMovieLikeCommandService;
 import lombok.RequiredArgsConstructor;
