+++ src/main/java/com/mingles/metamingle/shortform/query/application/controller/ShortFormQueryController.java
@@ -1,7 +1,7 @@
 package com.mingles.metamingle.shortform.query.application.controller;
 
-import com.mingles.metamingle.auth.JwtTokenProvider;
-import com.mingles.metamingle.common.ApiResponse;
+import com.mingles.metamingle.global.auth.JwtTokenProvider;
+import com.mingles.metamingle.global.common.ApiResponse;
 import com.mingles.metamingle.shortform.query.application.dto.response.GetShortFormListResponse;
 import com.mingles.metamingle.shortform.query.application.dto.response.GetShortFormResponse;
 import com.mingles.metamingle.shortform.query.application.service.ShortFormQueryService;
